refactor: broken down settings client page
