+++ types/base.ts
@@ -32,12 +32,6 @@ export interface Team {
   updatedAt: Date;
 }
 
-export interface TeamSettings {
-  isPrivate: boolean;
-  allowMemberInvites: boolean;
-  requireApproval: boolean;
-}
-
 export interface Project {
   id: string;
   name: string;
