+++ types/forms/team.ts
@@ -7,3 +7,26 @@ export interface InviteFormMember {
   avatarUrl?: string | null;
   isExistingUser: boolean;
 }
+
+export interface PendingTeamMember {
+  id?: string;
+  email: string;
+  firstName?: string;
+  lastName?: string;
+  username?: string;
+  avatarUrl?: string;
+  isExistingUser?: boolean;
+}
+
+export interface FormData {
+  name: string;
+  slug: string;
+  description: string;
+  settings: TeamSettings;
+}
+
+export interface TeamSettings {
+  isPrivate: boolean;
+  allowMemberInvites: boolean;
+  requireApproval: boolean;
+}
